· 移除Sandbox，创建画廊作为迁移工作的展示。
