+++ OpenHarmony.Avalonia.sln
@@ -15,9 +15,29 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AvaloniaApp.Desktop", "Src\
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenHarmony.NDK.Bindings", "ThirdParty\OpenHarmony.NDK.Bindings\OpenHarmony.NDK.Bindings\OpenHarmony.NDK.Bindings.csproj", "{F8E2F26E-DAE3-4CA7-924E-7F05E655E71D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SandBox", "Src\Example\SandBox\SandBox.csproj", "{0DB36DC2-D05D-46F8-9C89-FB15778DA2CA}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AOOH_Gallery", "Src\Example\AOOH_Gallery\AOOH_Gallery.csproj", "{0DB36DC2-D05D-46F8-9C89-FB15778DA2CA}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SandBox.Desktop", "Src\Example\SandBox.Desktop\SandBox.Desktop.csproj", "{A43F296F-D6DD-4F64-8F35-2622AF619254}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AOOH_Gallery.Desktop", "Src\Example\AOOH_Gallery.Desktop\AOOH_Gallery.Desktop.csproj", "{A43F296F-D6DD-4F64-8F35-2622AF619254}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "AOOH_Gallery", "AOOH_Gallery", "{492440ED-F090-4FFE-918F-20FD73E6DE4C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Semi.Avalonia.Demo", "ThirdParty\Semi.Avalonia\demo\Semi.Avalonia.Demo\Semi.Avalonia.Demo.csproj", "{0B866685-11D8-41D3-8BAA-A0D176358AC3}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ursa.Demo", "ThirdParty\Ursa.Avalonia\demo\Ursa.Demo\Ursa.Demo.csproj", "{ABD9CA64-B006-47A8-848A-21D43393DE73}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Irihitech", "Irihitech", "{7812D870-170F-488A-A058-F8C1433C178A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Semi.Avalonia", "ThirdParty\Semi.Avalonia\src\Semi.Avalonia\Semi.Avalonia.csproj", "{86579D70-3255-433F-B7C7-9F17EB63A655}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Semi.Avalonia.ColorPicker", "ThirdParty\Semi.Avalonia\src\Semi.Avalonia.ColorPicker\Semi.Avalonia.ColorPicker.csproj", "{0B4EA84D-3379-4530-86A9-36416CE38583}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Semi.Avalonia.TreeDataGrid", "ThirdParty\Semi.Avalonia\src\Semi.Avalonia.TreeDataGrid\Semi.Avalonia.TreeDataGrid.csproj", "{D3D1AB90-E346-4B83-8609-5789426E8C52}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Semi.Avalonia.DataGrid", "ThirdParty\Semi.Avalonia\src\Semi.Avalonia.DataGrid\Semi.Avalonia.DataGrid.csproj", "{39F1FF7A-3EE4-4AB8-A6A2-FA36AB703490}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ursa", "ThirdParty\Ursa.Avalonia\src\Ursa\Ursa.csproj", "{E8623BDC-FA99-4D07-8227-0152AC186377}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ursa.Themes.Semi", "ThirdParty\Ursa.Avalonia\src\Ursa.Themes.Semi\Ursa.Themes.Semi.csproj", "{51CD02F7-9D11-49C6-A1D0-8AC9BAABF505}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -53,6 +73,38 @@ Global
 		{A43F296F-D6DD-4F64-8F35-2622AF619254}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A43F296F-D6DD-4F64-8F35-2622AF619254}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A43F296F-D6DD-4F64-8F35-2622AF619254}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0B866685-11D8-41D3-8BAA-A0D176358AC3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0B866685-11D8-41D3-8BAA-A0D176358AC3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0B866685-11D8-41D3-8BAA-A0D176358AC3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0B866685-11D8-41D3-8BAA-A0D176358AC3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{ABD9CA64-B006-47A8-848A-21D43393DE73}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{ABD9CA64-B006-47A8-848A-21D43393DE73}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{ABD9CA64-B006-47A8-848A-21D43393DE73}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{ABD9CA64-B006-47A8-848A-21D43393DE73}.Release|Any CPU.Build.0 = Release|Any CPU
+		{86579D70-3255-433F-B7C7-9F17EB63A655}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{86579D70-3255-433F-B7C7-9F17EB63A655}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{86579D70-3255-433F-B7C7-9F17EB63A655}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{86579D70-3255-433F-B7C7-9F17EB63A655}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0B4EA84D-3379-4530-86A9-36416CE38583}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0B4EA84D-3379-4530-86A9-36416CE38583}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0B4EA84D-3379-4530-86A9-36416CE38583}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0B4EA84D-3379-4530-86A9-36416CE38583}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D3D1AB90-E346-4B83-8609-5789426E8C52}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D3D1AB90-E346-4B83-8609-5789426E8C52}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D3D1AB90-E346-4B83-8609-5789426E8C52}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D3D1AB90-E346-4B83-8609-5789426E8C52}.Release|Any CPU.Build.0 = Release|Any CPU
+		{39F1FF7A-3EE4-4AB8-A6A2-FA36AB703490}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{39F1FF7A-3EE4-4AB8-A6A2-FA36AB703490}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{39F1FF7A-3EE4-4AB8-A6A2-FA36AB703490}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{39F1FF7A-3EE4-4AB8-A6A2-FA36AB703490}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E8623BDC-FA99-4D07-8227-0152AC186377}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E8623BDC-FA99-4D07-8227-0152AC186377}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E8623BDC-FA99-4D07-8227-0152AC186377}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E8623BDC-FA99-4D07-8227-0152AC186377}.Release|Any CPU.Build.0 = Release|Any CPU
+		{51CD02F7-9D11-49C6-A1D0-8AC9BAABF505}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{51CD02F7-9D11-49C6-A1D0-8AC9BAABF505}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{51CD02F7-9D11-49C6-A1D0-8AC9BAABF505}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{51CD02F7-9D11-49C6-A1D0-8AC9BAABF505}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -60,7 +112,17 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{32CF34B5-F75A-4A00-A730-90140C0A792D} = {F1FEDF1B-E967-4070-B986-C11E9D6D6C61}
 		{D1AC0C3E-A28F-4604-9B04-74101768F712} = {F1FEDF1B-E967-4070-B986-C11E9D6D6C61}
-		{0DB36DC2-D05D-46F8-9C89-FB15778DA2CA} = {F1FEDF1B-E967-4070-B986-C11E9D6D6C61}
-		{A43F296F-D6DD-4F64-8F35-2622AF619254} = {F1FEDF1B-E967-4070-B986-C11E9D6D6C61}
+		{492440ED-F090-4FFE-918F-20FD73E6DE4C} = {F1FEDF1B-E967-4070-B986-C11E9D6D6C61}
+		{0DB36DC2-D05D-46F8-9C89-FB15778DA2CA} = {492440ED-F090-4FFE-918F-20FD73E6DE4C}
+		{A43F296F-D6DD-4F64-8F35-2622AF619254} = {492440ED-F090-4FFE-918F-20FD73E6DE4C}
+		{7812D870-170F-488A-A058-F8C1433C178A} = {F1FEDF1B-E967-4070-B986-C11E9D6D6C61}
+		{0B866685-11D8-41D3-8BAA-A0D176358AC3} = {7812D870-170F-488A-A058-F8C1433C178A}
+		{ABD9CA64-B006-47A8-848A-21D43393DE73} = {7812D870-170F-488A-A058-F8C1433C178A}
+		{86579D70-3255-433F-B7C7-9F17EB63A655} = {7812D870-170F-488A-A058-F8C1433C178A}
+		{0B4EA84D-3379-4530-86A9-36416CE38583} = {7812D870-170F-488A-A058-F8C1433C178A}
+		{D3D1AB90-E346-4B83-8609-5789426E8C52} = {7812D870-170F-488A-A058-F8C1433C178A}
+		{39F1FF7A-3EE4-4AB8-A6A2-FA36AB703490} = {7812D870-170F-488A-A058-F8C1433C178A}
+		{E8623BDC-FA99-4D07-8227-0152AC186377} = {7812D870-170F-488A-A058-F8C1433C178A}
+		{51CD02F7-9D11-49C6-A1D0-8AC9BAABF505} = {7812D870-170F-488A-A058-F8C1433C178A}
 	EndGlobalSection
 EndGlobal
