+++ Src/Entry/Entry.csproj
@@ -13,7 +13,7 @@
         <ProjectReference Include="..\..\ThirdParty\OpenHarmony.NDK.Bindings\OpenHarmony.NDK.Bindings\OpenHarmony.NDK.Bindings.csproj"/>
         <ProjectReference Include="..\Avalonia.OpenHarmony\Avalonia.OpenHarmony.csproj"/>
         <ProjectReference Include="..\Example\AvaloniaApp\AvaloniaApp.csproj"/>
-        <ProjectReference Include="..\Example\SandBox\SandBox.csproj"/>
+        <ProjectReference Include="..\Example\AOOH_Gallery\AOOH_Gallery.csproj" />
     </ItemGroup>
     <ItemGroup>
         <RdXmlFile Include="rd.xml"/>
