+++ Src/Entry/Properties/PublishProfiles/PublishAmd64.pubxml
@@ -4,7 +4,7 @@ https://go.microsoft.com/fwlink/?LinkID=208121.
 -->
 <Project>
     <PropertyGroup>
-        <Configuration>Debug</Configuration>
+        <Configuration>Releases</Configuration>
         <Platform>Any CPU</Platform>
         <PublishDir>../../OHOS_Project/entry/libs/x86_64</PublishDir>
         <PublishProtocol>FileSystem</PublishProtocol>
