+++ Src/Example/AOOH_Gallery.Desktop/AOOH_Gallery.Desktop.csproj
@@ -1,0 +1,19 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>net9.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <ApplicationManifest>app.manifest</ApplicationManifest>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Avalonia.Desktop" Version="11.2.8"/>
+    </ItemGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\AOOH_Gallery\AOOH_Gallery.csproj" />
+    </ItemGroup>
+
+</Project>
