+++ Src/Example/AOOH_Gallery/App.axaml.cs
@@ -1,0 +1,31 @@
+using AOOH_Gallery.ViewModels;
+using AOOH_Gallery.Views;
+using Avalonia;
+using Avalonia.Controls.ApplicationLifetimes;
+using Avalonia.Markup.Xaml;
+
+namespace AOOH_Gallery;
+
+public class App : Application
+{
+    public override void Initialize()
+    {
+        AvaloniaXamlLoader.Load(this);
+    }
+
+    public override void OnFrameworkInitializationCompleted()
+    {
+        if (ApplicationLifetime is IClassicDesktopStyleApplicationLifetime desktop)
+            desktop.MainWindow = new MainWindow
+            {
+                DataContext = new MainViewViewModel()
+            };
+        else if (ApplicationLifetime is ISingleViewApplicationLifetime singleViewPlatform)
+            singleViewPlatform.MainView = new MainView
+            {
+                DataContext = null
+            };
+
+        base.OnFrameworkInitializationCompleted();
+    }
+}
