+++ Src/Example/AOOH_Gallery/ViewLocator.cs
@@ -1,0 +1,27 @@
+using System;
+using AOOH_Gallery.ViewModels;
+using Avalonia.Controls;
+using Avalonia.Controls.Templates;
+
+namespace AOOH_Gallery;
+
+public class ViewLocator : IDataTemplate
+{
+    public Control? Build(object? param)
+    {
+        if (param is null)
+            return null;
+
+        var name = param.GetType().FullName!.Replace("ViewModel", "View", StringComparison.Ordinal);
+        var type = Type.GetType(name);
+
+        if (type != null) return (Control)Activator.CreateInstance(type)!;
+
+        return new TextBlock { Text = "Not Found: " + name };
+    }
+
+    public bool Match(object? data)
+    {
+        return data is ViewModelBase;
+    }
+}
