+++ Src/Example/AOOH_Gallery/ViewModels/MainViewViewModel.cs
@@ -1,0 +1,6 @@
+﻿namespace AOOH_Gallery.ViewModels;
+
+public class MainViewViewModel : ViewModelBase
+{
+    public string Greeting { get; } = "Welcome to Avalonia!";
+}
