+++ Src/Example/AOOH_Gallery/ViewModels/ViewModelBase.cs
@@ -1,0 +1,7 @@
+﻿using ReactiveUI;
+
+namespace AOOH_Gallery.ViewModels;
+
+public class ViewModelBase : ReactiveObject
+{
+}
