+++ Src/Example/AOOH_Gallery/Views/MainWindow.axaml
@@ -1,0 +1,21 @@
+<Window xmlns="https://github.com/avaloniaui"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:vm="using:AOOH_Gallery.ViewModels"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:views="clr-namespace:AOOH_Gallery.Views"
+        mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
+        x:Class="AOOH_Gallery.Views.MainWindow"
+        x:DataType="vm:MainViewViewModel"
+        Icon="/Assets/avalonia-logo.ico"
+        Title="AOOH_Gallery">
+
+    <Design.DataContext>
+        <!-- This only sets the DataContext for the previewer in an IDE,
+             to set the actual DataContext for runtime, set the DataContext property in code (look at App.axaml.cs) -->
+        <vm:MainViewViewModel />
+    </Design.DataContext>
+
+    <views:MainView />
+
+</Window>
