+++ Src/Example/AOOH_Gallery/Views/StorgePage.axaml.cs
@@ -1,0 +1,17 @@
+using System;
+using System.IO;
+using System.Linq;
+using Avalonia.Controls;
+using Avalonia.Interactivity;
+using Avalonia.OpenHarmony;
+using Avalonia.Platform.Storage;
+
+namespace AOOH_Gallery.Views;
+
+public partial class StorgePage : UserControl
+{
+    public StorgePage()
+    {
+        InitializeComponent();
+    }
+}
