+++ Src/Example/AOOH_Gallery/Views/SystemFonts.axaml
@@ -1,0 +1,13 @@
+<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
+             x:Class="AOOH_Gallery.Views.SystemFonts">
+    <HeaderedContentControl Theme="{DynamicResource GroupBox}">
+        <HeaderedContentControl.Header>
+            你能够直接使用的系统字体如下
+        </HeaderedContentControl.Header>
+        <ListBox Name="ListBox" />
+    </HeaderedContentControl>
+</UserControl>
