+++ Src/Example/AOOH_Gallery/Views/SystemFonts.axaml.cs
@@ -1,0 +1,22 @@
+using System.Linq;
+using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Interactivity;
+using Avalonia.Markup.Xaml;
+using Avalonia.Media;
+
+namespace AOOH_Gallery.Views;
+
+public partial class SystemFonts : UserControl
+{
+    public SystemFonts()
+    {
+        InitializeComponent();
+    }
+
+    protected override void OnLoaded(RoutedEventArgs e)
+    {
+        base.OnLoaded(e);
+        ListBox.ItemsSource = FontManager.Current.SystemFonts.Select(x => x.Name);
+    }
+}
