--- Src/Example/SandBox.Desktop/SandBox.Desktop.csproj
@@ -1,19 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <OutputType>Exe</OutputType>
-        <TargetFramework>net9.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <ApplicationManifest>app.manifest</ApplicationManifest>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Avalonia.Desktop" Version="11.2.8"/>
-    </ItemGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\SandBox\SandBox.csproj"/>
-    </ItemGroup>
-
-</Project>
