--- Src/Example/SandBox.Desktop/app.manifest
@@ -1,18 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<assembly manifestVersion="1.0" xmlns="urn:schemas-microsoft-com:asm.v1">
-    <!-- This manifest is used on Windows only.
-         Don't remove it as it might cause problems with window transparency and embedded controls.
-         For more details visit https://learn.microsoft.com/en-us/windows/win32/sbscs/application-manifests -->
-    <assemblyIdentity version="1.0.0.0" name="SandBox.Desktop"/>
-
-    <compatibility xmlns="urn:schemas-microsoft-com:compatibility.v1">
-        <application>
-            <!-- A list of the Windows versions that this application has been tested on
-                 and is designed to work with. Uncomment the appropriate elements
-                 and Windows will automatically select the most compatible environment. -->
-
-            <!-- Windows 10 -->
-            <supportedOS Id="{8e0f7a12-bfb3-4fe8-b9a5-48fd50a15a9a}"/>
-        </application>
-    </compatibility>
-</assembly>
