--- Src/Example/SandBox/SandBox.csproj
@@ -1,46 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-    <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
-        <Nullable>enable</Nullable>
-        <BuiltInComInteropSupport>true</BuiltInComInteropSupport>
-        <AvaloniaUseCompiledBindingsByDefault>true</AvaloniaUseCompiledBindingsByDefault>
-        <LangVersion>preview</LangVersion>
-        <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <Folder Include="Models\"/>
-        <AvaloniaResource Include="Assets\**"/>
-    </ItemGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Avalonia" Version="11.2.8"/>
-        <PackageReference Include="Avalonia.Themes.Fluent" Version="11.2.8"/>
-        <PackageReference Include="Avalonia.Fonts.Inter" Version="11.2.8"/>
-        <!--Condition below is needed to remove Avalonia.Diagnostics package from build output in Release configuration.-->
-        <PackageReference Include="Avalonia.Diagnostics" Version="11.2.8">
-            <IncludeAssets Condition="'$(Configuration)' != 'Debug'">None</IncludeAssets>
-            <PrivateAssets Condition="'$(Configuration)' != 'Debug'">All</PrivateAssets>
-        </PackageReference>
-        <PackageReference Include="Avalonia.ReactiveUI" Version="11.2.8"/>
-        <PackageReference Include="Semi.Avalonia" Version="11.2.1.6"/>
-        <PackageReference Include="Semi.Avalonia.TreeDataGrid" Version="11.0.10.2"/>
-        <PackageReference Include="Irihi.Ursa" Version="1.10.0"/>
-        <PackageReference Include="Irihi.Ursa.Themes.Semi" Version="1.10.0"/>
-        <PackageReference Include="ReactiveUI.SourceGenerators" Version="2.1.27">
-            <PrivateAssets>all</PrivateAssets>
-            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-        </PackageReference>
-        <PackageReference Include="ReactiveMarbles.ObservableEvents.SourceGenerator" Version="1.3.1"/>
-    </ItemGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\..\Avalonia.OpenHarmony\Avalonia.OpenHarmony.csproj"/>
-    </ItemGroup>
-
-    <ItemGroup>
-      <Reference Include="SkiaSharp">
-        <HintPath>..\..\Avalonia.OpenHarmony\libs\SkiaSharp\SkiaSharp.dll</HintPath>
-      </Reference>
-    </ItemGroup>
-</Project>
