--- Src/Example/SandBox/ViewModels/MainViewViewModel.cs
@@ -1,6 +1,0 @@
-﻿namespace SandBox.ViewModels;
-
-public class MainViewViewModel : ViewModelBase
-{
-    public string Greeting { get; } = "Welcome to Avalonia!";
-}
