--- Src/Example/SandBox/ViewModels/ViewModelBase.cs
@@ -1,7 +1,0 @@
-﻿using ReactiveUI;
-
-namespace SandBox.ViewModels;
-
-public class ViewModelBase : ReactiveObject
-{
-}
