--- Src/Example/SandBox/Views/MainView.axaml
@@ -1,33 +1,0 @@
-﻿<UserControl xmlns="https://github.com/avaloniaui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:openHarmony="clr-namespace:Avalonia.OpenHarmony;assembly=Avalonia.OpenHarmony"
-             xmlns:views="clr-namespace:SandBox.Views"
-             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             x:Class="SandBox.Views.MainView">
-    <Grid
-        Margin="50"
-        RowDefinitions="55*,45*">
-        <TabControl TabStripPlacement="Left">
-            <TabItem Header="输入法">
-                <views:IMEPage />
-            </TabItem>
-            <TabItem Header="文件存储服务">
-                <views:StorgePage />
-            </TabItem>
-        </TabControl>
-        <ListBox
-            Grid.Row="1"
-            Name="ListBox"
-            ItemsSource="{x:Static openHarmony:OHDebugHelper.Logs}">
-            <ListBox.ItemTemplate>
-                <DataTemplate>
-                    <Border Theme="{DynamicResource CardBorder}">
-                        <TextBlock Text="{Binding Message}" />
-                    </Border>
-                </DataTemplate>
-            </ListBox.ItemTemplate>
-        </ListBox>
-    </Grid>
-</UserControl>
