--- Src/Example/SandBox/Views/MainWindow.axaml.cs
@@ -1,11 +1,0 @@
-using Avalonia.Controls;
-
-namespace SandBox.Views;
-
-public partial class MainWindow : Window
-{
-    public MainWindow()
-    {
-        InitializeComponent();
-    }
-}
